+++ src/headless/store/services/social-sharing-service.ts
@@ -6,8 +6,8 @@ import {
 import { SignalsServiceDefinition } from '@wix/services-definitions/core-services/signals';
 import type { Signal } from '../../Signal';
 import {
-  SOCIAL_PLATFORM_CONFIGS,
   SocialPlatform,
+  SocialPlatformShareUrl,
 } from '../enums/social-platform-enums';
 
 export interface SharingPlatform {
@@ -47,12 +47,38 @@ export const SocialSharingService = implementService.withConfig<{
 }>()(SocialSharingServiceDefinition, ({ getService }) => {
   const signalsService = getService(SignalsServiceDefinition);
 
+    // Platform metadata is handled in components layer, only business logic here
   const availablePlatforms: Signal<SharingPlatform[]> = signalsService.signal([
-    SOCIAL_PLATFORM_CONFIGS[SocialPlatform.FACEBOOK],
-    SOCIAL_PLATFORM_CONFIGS[SocialPlatform.TWITTER],
-    SOCIAL_PLATFORM_CONFIGS[SocialPlatform.LINKEDIN],
-    SOCIAL_PLATFORM_CONFIGS[SocialPlatform.WHATSAPP],
-    SOCIAL_PLATFORM_CONFIGS[SocialPlatform.EMAIL],
+    {
+      name: 'Facebook',
+      icon: 'facebook',
+      color: '#1877F2',
+      shareUrl: SocialPlatformShareUrl.FACEBOOK as string,
+    },
+    {
+      name: 'Twitter',
+      icon: 'twitter', 
+      color: '#1DA1F2',
+      shareUrl: SocialPlatformShareUrl.TWITTER as string,
+    },
+    {
+      name: 'LinkedIn',
+      icon: 'linkedin',
+      color: '#0A66C2', 
+      shareUrl: SocialPlatformShareUrl.LINKEDIN as string,
+    },
+    {
+      name: 'WhatsApp',
+      icon: 'whatsapp',
+      color: '#25D366',
+      shareUrl: SocialPlatformShareUrl.WHATSAPP as string,
+    },
+    {
+      name: 'Email',
+      icon: 'mail',
+      color: '#EA4335',
+      shareUrl: SocialPlatformShareUrl.EMAIL as string,
+    },
   ]);
 
   const shareCount: Signal<number> = signalsService.signal(0 as any);
@@ -80,9 +106,7 @@ export const SocialSharingService = implementService.withConfig<{
     title: string,
     description?: string
   ) => {
-    const shareUrl = new URL(
-      SOCIAL_PLATFORM_CONFIGS[SocialPlatform.FACEBOOK].shareUrl!
-    );
+    const shareUrl = new URL(SocialPlatformShareUrl.FACEBOOK);
     shareUrl.searchParams.set('u', url);
     shareUrl.searchParams.set(
       'quote',
@@ -93,9 +117,7 @@ export const SocialSharingService = implementService.withConfig<{
   };
 
   const shareToTwitter = (url: string, text: string, hashtags?: string[]) => {
-    const shareUrl = new URL(
-      SOCIAL_PLATFORM_CONFIGS[SocialPlatform.TWITTER].shareUrl!
-    );
+    const shareUrl = new URL(SocialPlatformShareUrl.TWITTER);
     shareUrl.searchParams.set('url', url);
     shareUrl.searchParams.set('text', text);
     if (hashtags && hashtags.length > 0) {
@@ -106,9 +128,7 @@ export const SocialSharingService = implementService.withConfig<{
   };
 
   const shareToLinkedIn = (url: string, title: string, summary?: string) => {
-    const shareUrl = new URL(
-      SOCIAL_PLATFORM_CONFIGS[SocialPlatform.LINKEDIN].shareUrl!
-    );
+    const shareUrl = new URL(SocialPlatformShareUrl.LINKEDIN);
     shareUrl.searchParams.set('url', url);
     shareUrl.searchParams.set('title', title);
     if (summary) {
@@ -120,14 +140,14 @@ export const SocialSharingService = implementService.withConfig<{
 
   const shareToWhatsApp = (url: string, text: string) => {
     const message = `${text} ${url}`;
-    const shareUrl = `${SOCIAL_PLATFORM_CONFIGS[SocialPlatform.WHATSAPP].shareUrl}?text=${encodeURIComponent(message)}`;
+    const shareUrl = `${SocialPlatformShareUrl.WHATSAPP}?text=${encodeURIComponent(message)}`;
 
     openShareWindow(shareUrl, SocialPlatform.WHATSAPP);
   };
 
   const shareToEmail = (url: string, subject: string, body: string) => {
     const emailBody = `${body}\n\n${url}`;
-    const mailtoUrl = `${SOCIAL_PLATFORM_CONFIGS[SocialPlatform.EMAIL].shareUrl}?subject=${encodeURIComponent(
+    const mailtoUrl = `${SocialPlatformShareUrl.EMAIL}?subject=${encodeURIComponent(
       subject
     )}&body=${encodeURIComponent(emailBody)}`;
